+++ crates/proto/src/domain/note.rs
@@ -1,6 +1,7 @@
 use miden_objects::{
     Digest, Felt,
-    note::{NoteExecutionHint, NoteId, NoteInclusionProof, NoteMetadata, NoteTag, NoteType},
+    note::{Note, NoteExecutionHint, NoteId, NoteInclusionProof, NoteMetadata, NoteTag, NoteType},
+    utils::Serializable,
 };
 
 use crate::{
@@ -24,6 +25,12 @@ impl TryFrom<proto::NoteMetadata> for NoteMetadata {
         let aux = Felt::try_from(value.aux).map_err(|_| ConversionError::NotAValidFelt)?;
 
         Ok(NoteMetadata::new(sender, note_type, tag, execution_hint, aux)?)
+    }
+}
+
+impl From<Note> for proto::NetworkNote {
+    fn from(note: Note) -> Self {
+        Self { note: note.to_bytes() }
     }
 }